# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberShield AI - Cybersecurity Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: 'USER' | 'ADMIN';
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: UserRole;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ProfileUpdate {
  /** @minLength 2 */
  fullName?: string;
}

export interface PasswordChange {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface MessageResponse {
  message: string;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  content: string;
  /** @nullable */
  videoUrl?: string | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type CourseDetailUserProgress = {
  /** @nullable */
  score?: number | null;
  completed?: boolean;
  /** @nullable */
  completedAt?: string | null;
} | null;

export interface CourseDetail {
  id: number;
  title: string;
  description: string;
  content: string;
  /** @nullable */
  videoUrl?: string | null;
  createdAt: string;
  quizCount: number;
  /** @nullable */
  userProgress?: CourseDetailUserProgress;
}

export interface CourseInput {
  /** @minLength 3 */
  title: string;
  description: string;
  content: string;
  /** @nullable */
  videoUrl?: string | null;
}

export interface CourseUpdate {
  title?: string;
  description?: string;
  content?: string;
  /** @nullable */
  videoUrl?: string | null;
}

export type QuizCorrectAnswer = typeof QuizCorrectAnswer[keyof typeof QuizCorrectAnswer];


export const QuizCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface Quiz {
  id: number;
  courseId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: QuizCorrectAnswer;
  createdAt: string;
}

export type QuizInputCorrectAnswer = typeof QuizInputCorrectAnswer[keyof typeof QuizInputCorrectAnswer];


export const QuizInputCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface QuizInput {
  courseId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: QuizInputCorrectAnswer;
}

export type QuizUpdateCorrectAnswer = typeof QuizUpdateCorrectAnswer[keyof typeof QuizUpdateCorrectAnswer];


export const QuizUpdateCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export interface QuizUpdate {
  question?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: QuizUpdateCorrectAnswer;
}

export type QuizSubmissionAnswersItemAnswer = typeof QuizSubmissionAnswersItemAnswer[keyof typeof QuizSubmissionAnswersItemAnswer];


export const QuizSubmissionAnswersItemAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;

export type QuizSubmissionAnswersItem = {
  quizId: number;
  answer: QuizSubmissionAnswersItemAnswer;
};

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
}

export interface Progress {
  id: number;
  userId: number;
  courseId: number;
  /** @nullable */
  score: number | null;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  course?: Course;
}

export interface QuizResult {
  score: number;
  total: number;
  passed: boolean;
  progress: Progress;
}

export type ThreatReportRiskLevel = typeof ThreatReportRiskLevel[keyof typeof ThreatReportRiskLevel];


export const ThreatReportRiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type ThreatReportStatus = typeof ThreatReportStatus[keyof typeof ThreatReportStatus];


export const ThreatReportStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  RESOLVED: 'RESOLVED',
} as const;

export interface ThreatReport {
  id: number;
  userId: number;
  title: string;
  description: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  aiAnalysis: string;
  riskLevel: ThreatReportRiskLevel;
  status: ThreatReportStatus;
  createdAt: string;
}

export type ThreatReportWithUserRiskLevel = typeof ThreatReportWithUserRiskLevel[keyof typeof ThreatReportWithUserRiskLevel];


export const ThreatReportWithUserRiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export type ThreatReportWithUserStatus = typeof ThreatReportWithUserStatus[keyof typeof ThreatReportWithUserStatus];


export const ThreatReportWithUserStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  RESOLVED: 'RESOLVED',
} as const;

export type ThreatReportWithUserUser = {
  id: number;
  fullName: string;
  email: string;
};

export interface ThreatReportWithUser {
  id: number;
  userId: number;
  title: string;
  description: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  aiAnalysis: string;
  riskLevel: ThreatReportWithUserRiskLevel;
  status: ThreatReportWithUserStatus;
  createdAt: string;
  user: ThreatReportWithUserUser;
}

export interface ThreatInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type ThreatStatusUpdateStatus = typeof ThreatStatusUpdateStatus[keyof typeof ThreatStatusUpdateStatus];


export const ThreatStatusUpdateStatus = {
  PENDING: 'PENDING',
  REVIEWED: 'REVIEWED',
  RESOLVED: 'RESOLVED',
} as const;

export interface ThreatStatusUpdate {
  status: ThreatStatusUpdateStatus;
}

export interface Announcement {
  id: number;
  title: string;
  message: string;
  createdAt: string;
}

export interface AnnouncementInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 5 */
  message: string;
}

export interface AdminStats {
  totalUsers: number;
  totalCourses: number;
  totalThreats: number;
  pendingThreats: number;
  resolvedThreats: number;
  avgQuizScore: number;
  completedCourses: number;
}

export interface RecentActivity {
  recentThreats: ThreatReportWithUser[];
  recentUsers: User[];
  recentProgress: Progress[];
}