# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberShield AI - Cybersecurity Awareness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or email already used
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS (admin)
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: All courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by id
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # QUIZZES
  /quizzes/course/{courseId}:
    get:
      operationId: listQuizzesByCourse
      tags: [quizzes]
      summary: Get all quiz questions for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{announcementId}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement (admin only)
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN DASHBOARD
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # User
    User:
      type: object
      required: [id, fullName, email, role, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time

    # Course
    Course:
      type: object
      required: [id, title, description, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CourseDetail:
      type: object
      required: [id, title, description, content, createdAt, quizCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        quizCount:
          type: integer
        userProgress:
          type: ["object", "null"]
          properties:
            score:
              type: ["integer", "null"]
            completed:
              type: boolean
            completedAt:
              type: ["string", "null"]
              format: date-time
    CourseInput:
      type: object
      required: [title, description, content]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: ["string", "null"]
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: ["string", "null"]

    # Quiz
    Quiz:
      type: object
      required: [id, courseId, question, optionA, optionB, optionC, optionD, correctAnswer, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
        createdAt:
          type: string
          format: date-time
...
    Announcement:
      type: object
      required: [id, title, message, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
          minLength: 3
        message:
          type: string
          minLength: 5

    # Admin
...