# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberShield AI - Cybersecurity Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizCorrectAnswer } from './quizCorrectAnswer';

export interface Quiz {
  id: number;
  courseId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: QuizCorrectAnswer;
  createdAt: Date;
}